Adjust dependency trees for gitbook sync (Recidiviz/recidiviz-data#7600)
The dependency trees included in the views section for gitbook at times exceed the apparent line limit of ~525. To fix this, we check the tree length before including it in the markdown. If the tree string is too long, we provide a script for printing out the tree to console, rather than displaying the tree in markdown.

GitOrigin-RevId: 674835eb840816dd4304c20d3cd2f9c0c88a0081