[Entity Matching] Downgrade assert to warning for enum-like objects (Recidiviz/recidiviz-data#7809)
This PR downgrades the non_placeholder_ingest_types inclusion assert to a warning if we discover a non-placeholder enum-like object (e.g. StatePersonRace, StatePersonEthnicity etc) that isn't in the set. These enum-like objects are not actually ever involved in the post-processing that uses the non_placeholder_ingest_types set, and are more likely to be not hydrated for a sizeable chunk of objects.

Work towards Recidiviz/recidiviz-data#6118.

GitOrigin-RevId: 28e13ce0bb2e6cf6a02f8410b8c6c4827f359568