Update days in dataflow_metrics dataset limit (Recidiviz/recidiviz-data#7853)
Recently our supervision_population_metrics table has grown to 1.67 TB in staging, causing the materialization of the most_recent_supervision_population_metrics view (which happens multiple times a day) to be quite expensive. Keeping 7-days worth of data in our dataflow_metrics dataset is super, super conservative, considering the logic in the move_old_dataflow_metrics_to_cold_storage function that keeps metrics around if they are from "the most recent job_id for a metric in a state where that metric is regularly calculated". This PR reduces the MAX_DAYS_IN_DATAFLOW_METRICS_TABLE from 7 to 2, which should significantly reduce the size of the dataflow_metrics tables.

As we onboard more states we will have to reconsider how we store dataflow output, with potentially state-segmented dataflow output datasets. But for now, this is a very low effort high reward win.

GitOrigin-RevId: 01cdb9884396d2afadae701eac7e9f6a4cf19587