Script for running CloudSQL -> BQ refresh to sandbox (Recidiviz/recidiviz-data#7982)
Adds a script for loading the current state of a Postgres instance into a sandbox BigQuery dataset. Can be used to test changes to the CloudSQL -> BigQuery refresh. Or when debugging issues in Postgres when the data has not yet been loaded to BigQuery (either because refresh for a given region is paused, or the daily refresh just hasn't run yet).

I've been using this code in a less-refined form to test changes to the federated export and figured it should be a script for everyone!

Related to Recidiviz/recidiviz-data#6067.

GitOrigin-RevId: 4fde7fe8e9f010b6bdfe033657374e08121ce255