Full start reason coverage in sessions admission dedup view (Recidiviz/recidiviz-data#7913)
My changes to incarceration periods are starting to put SANCTION_ADMISSION admission reasons on periods of admission from supervision due to a sanction (for treatment or shock incarceration). I went to see how this value was ranked in the sessions start de-duping and noticed it wasn't in there yet! So, updated that view to ensure we're always covering all possible enum values.

GitOrigin-RevId: 1e439b9be09080b9505fa71591097314ee4c8bb8