[Pseduo locks] Update no_active_locks_with_prefix check to ingore expired locks (Recidiviz/recidiviz-data#8049)
Fixes a bug in our pseudo-locking code where no_active_locks_with_prefix() was returning false if it found EXPIRED locks with a given prefix.

Also changes some naming related to the lock contents, since we were using the term contents to refer to both the full lock text and also the payload that was included along with metadata. Changed so "body" means the whole thing and "payload" means the user-supplied part.

GitOrigin-RevId: 1d0c69fb2ab86844377cf354095155f2ba0ed1fc