[Tools] Add file tag filter to import_raw_files_to_sandbox script (Recidiviz/recidiviz-data#8365)
I keep hacking this script locally to filter file tags when I only want to test if one or two files in a dump parse. This change adds an option to only import files with certain tags, even if more files exist in the bucket.

Built this for Recidiviz/recidiviz-data#8362

GitOrigin-RevId: d14fa3d94c7e2aab3848f4ae22cb1dd3f6dc4b06