[Multi-DB] Update scheduler to only consider tasks from the same instance (Recidiviz/recidiviz-data#8647)
In the controller, we check process_job_queue_info.size() and not just_finished_job to abort because there is already another process job request in progress. However, this queue is now shared between primary and secondary, so a primary job can block a secondary job.

This PR updates the logic to only look at tasks for a particular instance.

Closes Recidiviz/recidiviz-data#8636

GitOrigin-RevId: 7388772c00905ae76c0434a125a656daca5d11e3