[BQ refresh] Add federated refresh datasets to managed views list (Recidiviz/recidiviz-data#8684)
Some of the views in the federated refresh aren't actually managed by the deploy so we forgot to add them to the DATASETS_THAT_HAVE_EVER_BEEN_MANAGED list. This PR adds them and also adds a test so it will be hard to forget to add new ones down the line.

Closes Recidiviz/recidiviz-data#8675

GitOrigin-RevId: 2f76ba8ecec76816ae51dec18c42daa7b9ceb1ac