Updating lantern views to handle new `agent_external_id` values (Recidiviz/recidiviz-data#8656)
View updates for Lantern to handle the upcoming changes to agent_external_id in Recidiviz/recidiviz-data#8605.

This PR includes a temporary update to the agent_external_id_to_full_names to include a temporary agent_external_id_with_full_name column with the ID: FULL NAME format. (The view includes null handling and formatting needed in multiple views.) This field is then used for all of the officer fields in the lantern views, so that Recidiviz/recidiviz-data#8605 can land without us needing any downtime on the FE.

The Lantern dashboard needs to have a unique identifier for the officers for the "Officer" chart (supported by revocations_matrix_distribution_by_officer view). This change keeps the value in the officer column unique, which will be true before and after Recidiviz/recidiviz-data#8605 lands (and before and after the source of that column is changed from the temp agent_external_id_to_full_names back to agent_external_id).

The Lantern dashboard also needs the full names of officers for the filtered caseload table (revocations_matrix_filtered_caseload). This view will now start receiving a new officer_full_name column as well, and the FE should be updated to rely on that column for the officer names in the filtered caseload.

GitOrigin-RevId: 89456be3d1ce6ac42b71f1448fb8a1ebf23d8b92