[Multi-DB] Migrate ND PRIMARY DB in PROD (Recidiviz/recidiviz-data#8596)
Transitions ND primary DB to us_nd_primary in production. Since this is the last state to migrate in production and all other states will have been migrated in staging by the time this lands, we can simplify the logic to default to SQLAlchemyStateDatabaseVersion.PRIMARY for all states / projects.

More thorough cleanup to come in Recidiviz/recidiviz-data#7984.

Work towards Recidiviz/recidiviz-data#6226 (will close when flashing is actually complete for ND).

GitOrigin-RevId: 5bf7f590a41f13b474a3278f94eca8abcb034575