[Direct ingest] Improvements to move_storage_files_to_deprecated script (Recidiviz/recidiviz-data#8836)
This PR updates the move_storage_files_to_deprecated script to use the OperateOnStorageIngestFilesController which allows us to delete a bunch of semi-duplicate code and gives us split-file handling for free.

In order to do this, I had to update the OperateOnStorageIngestFilesController to support file tag filters. I also changed how we fetched the list of candidate directories for significant perf wins.

Closes Recidiviz/recidiviz-data#3716.
Closes Recidiviz/recidiviz-data#4847.

GitOrigin-RevId: 837069102d2fd1f108d14114ce131dc5bbd24dc2