[Schema Cleanup] Deprecate revocation_type for US_ND (Recidiviz/recidiviz-data#9013)
Deprecates the revocation_type and revocation_type_raw_text values on the StateSupervisionViolationResponse for US_ND. Also stops putting REVOCATION in the deprecated decision field on this entity. Creates new StateSupervisionViolationResponseDecisionEntry entities on these responses where the values that used to be in revocation_type are now stored in the decision field.

These values are not currently used anywhere in calculations or downstream analytics. I did not write migrations to create these new decision entry entities, and assume this will be covered by the next ND re-run.

GitOrigin-RevId: 09a6e78b3b5aaf713cd9131e54bb5ae7f01f96e9