[Bug Fix] Avoids cleaning up datasets for paused regions in federated refresh (Recidiviz/recidiviz-data#9097)
Refresh in prod was failing while the US_PA refresh was paused since it was cleaning up the associated us_pa_... datasets that aren't updated when a region is paused. This removes datasets associated with paused regions from the list of datasets that should be considered for clean up to avoid this issue.

GitOrigin-RevId: d94956bc8164a727760f288e00a5c9bee8b381f1