[Ingest mappings] Support for simple enum parsing in v2 mappings (Recidiviz/recidiviz-data#9081)
Adds support for simple enum mappings. Introduces the concept of an `EnumEntity`, an entity that just exists to wrap an enum in a list field, so that we can create objects like `StatePersonRace` in a less verbose way.

To review this PR:
1) Look at example test cases to get a sense of syntax
1) Look at Recidiviz/recidiviz-data#9082 to see how this plays out in a real ingest view
1) Review actual implementation.

Not in scope: enum mapper functions, support for sharing mappings across fields / files.

Closes Recidiviz/recidiviz-data#8977

GitOrigin-RevId: d43213007b087b7ea31bb067d2e3e9c80debf7a8