[Ingest mappings] Add deserializing factory for StateSupervisionPeriod (Recidiviz/recidiviz-data#9209)
Adds a deserializing factory for `StateSupervisionPeriod`. This transformation was relatively trivial with the exception of the `conditions` field. Previously, the IngestInfo object stored conditions as a list, only to concatenate during the conversion to the `entities.py` entities. This update pushes the responsibility of concatenating to the ingest controller (only actually applies to US_PA).

Part of Recidiviz/recidiviz-data#8909

GitOrigin-RevId: 8e957e8ca692075ba41b06dfac5b1f1489a9febe