[Schema Flattening] Move `StateIncarcerationIncident` to be on `StatePerson` (Recidiviz/recidiviz-data#9252)
Moves StateIncarcerationIncident to hang directly off of the StatePerson. Removes the relationship between StateIncarcerationIncident and StateIncarcerationPeriod.

Adds a new helper function for running pipeline tests so we don't have to manually list every single table required for a pipeline test. As we start adding more and more classes to the StatePerson these lists would have gotten very annoying to maintain.

GitOrigin-RevId: 95fafa2483b3e549f4d92a59a173cac78dda932b