[Schema cleanup] Add residency_status_raw_text field (Recidiviz/recidiviz-data#9299)
In the process of migrating a view that hydrates the `residency_status` field, I realized it has no corresponding raw text field which is a no go for the v2 ingest mappings framework.

This PR updates to add a `residency_status_raw_text` field to the schema which can be hydrated as we migrate to to v2 ingest mappings. It also removes the global converter override which wasn't configured correctly for the v2 mappings infrastructure. I actually think that it will be best to allow states to customize how we hydrate this field.

Work towards Recidiviz/recidiviz-data#8968.

GitOrigin-RevId: d31f88bef805f711175a6e40ac63baff1bea881d