[Performance] Cache Entity field types in CoreEntityFieldIndex (Recidiviz/recidiviz-data#9501)
In Recidiviz/recidiviz-data#9453 I updated our code to use a CoreEntityFieldIndex which cached the set of all fields of a given type for any DatabaseEntity class. This PR extends that to also cache results for Entity classes. I will rely on the performance of this functionality later in a fix to preserve functionality between ingest mappings v1 and v2 versions.

Work towards Recidiviz/recidiviz-data#9425.
Work towards Recidiviz/recidiviz-data#9356.

GitOrigin-RevId: d3a09ecb7a81893d200c8e6de5a5adf2ff0cc724