[Calc Entity Hydration] V2 Entity Hydration, and implementation for `program` pipeline  (Recidiviz/recidiviz-data#9844)
Implements the new form of entity hydration for the `program` pipeline. Also updates the `identifier` classes of other pipelines to receive new format of the dictionary that stores the entities after the extraction process.

The bulk of the changes and updates are in `extractor_utils.py`. This is where I've implemented the new extraction process. The new approach for entity hydration is:
- Have each pipeline define the set of entities required for that pipeline
- Query for // hydrate each entity in this list
- Determine the set of relationships that exist between the required entities
- Pull in the values required to hydrate these relationships (where necessary)
- Hydrate all relationships that exist between hydrated entities

This approach will give us efficiency wins for the pipelines where we are currently querying for//hydrating certain entity types multiple times, and also where we are bringing in many entities that aren't actually required for the pipeline (e.g. we're bringing in `StateSupervisionContact` and `StateAssessment` entities into the `violations` pipeline since they are connected to `StatePerson` even though they are completely unused by the pipeline).

The program pipeline as it stands is super simple, and doesn't hydrate related entities for anything other than `StatePerson`. For this reason, this change isn't an efficiency gain for the program pipeline (total shuffled data increased from 7.5 GB to 12 GB), but it is a huge win in code simplicity and having entity relationships auto-hydrated.

Have confirmed program pipeline output is identical with these changes!

GitOrigin-RevId: 607164a48c4008e19f1e537ce8bdd8372ad9ee72