[Entity Hydration] Hydrating reference tables along with entities (Recidiviz/recidiviz-data#9981)
This restructures when we read in person-level reference tables so that we can group all data with a person_id during a single CoGroupByKey function. This drastically reduces the amount of data that needs to get shuffled to bring in and connect all data relevant to a single person for the pipeline.

GitOrigin-RevId: 8e0ad48bc79aac342ae5e69f521622835d57d28d